Move drawer to its own folder
